Separate some test functions
Now system.rs is too long
