+++ src/racer/core.rs
@@ -5,7 +5,7 @@ use matchers::ImportInfo;
 use project_model::ProjectModelProvider;
 use rls_span;
 use std::cell::RefCell;
-use std::cmp::{max, min, Ordering};
+use std::cmp::Ordering;
 use std::collections::HashMap;
 use std::fs::File;
 use std::io;
@@ -13,7 +13,6 @@ use std::io::Read;
 use std::iter::{Fuse, Iterator};
 use std::ops::{Deref, Range};
 use std::rc::Rc;
-use std::slice;
 use std::{fmt, vec};
 use std::{path, str};
 use syntax::source_map;
