+++ tests/system.rs
@@ -3776,189 +3776,6 @@ fn finds_method_definition_in_1line_closure() {
     assert_eq!(got.matchstr, "push");
 }
 
-// issue 706
-mod trait_bounds {
-    use super::*;
-    #[test]
-    fn finds_type_parameter_for_fnarg() {
-        let src = "
-        fn main() {
-            trait Trait {
-                fn method(&self);
-            }
-            fn func<T: Trait>(arg: &T) {
-                arg.meth~
-            }
-        }
-        ";
-        assert_eq!(get_only_completion(src, None).matchstr, "method");
-    }
-
-    #[test]
-    fn completes_methods_for_fnarg_by_trait_bounds() {
-        let src = "
-        fn main() {
-            trait Trait {
-                fn method(&self);
-            }
-            fn func<T: Trait>(arg: &T) {
-                arg.meth~
-            }
-        }
-        ";
-        assert_eq!(get_only_completion(src, None).matchstr, "method");
-    }
-
-    #[test]
-    fn completes_methods_for_fnarg_by_where_clause() {
-        let src = "
-        fn main() {
-            trait Trait {
-                fn method(&self);
-            }
-            fn func<T>(arg: &T)
-            where
-                T: Trait,
-            {
-                arg.meth~
-            }
-        }
-        ";
-        assert_eq!(get_only_completion(src, None).matchstr, "method");
-    }
-
-    #[test]
-    fn completes_external_methods_for_fnarg_by_trait_bounds() {
-        let src = "
-        fn main() {
-            fn func<T: Debug + Clone>(arg: &T) {
-                arg.clo~
-            }
-        }
-        ";
-        assert!(
-            get_all_completions(src, None)
-                .into_iter()
-                .any(|ma| ma.matchstr == "clone")
-        );
-    }
-
-    #[test]
-    fn completes_inherited_methods_for_fnarg_by_trait_bounds() {
-        let src = "
-        fn main() {
-            trait Inherited {
-                fn inherited(&self);
-            }
-            trait Trait: Inherited {
-                fn method(&self);
-            }
-            fn func<T: Trait>(arg: &T) {
-                arg.inheri~
-            }
-        }
-        ";
-        assert_eq!(get_only_completion(src, None).matchstr, "inherited");
-    }
-
-    // test for checking racer don't cause INF loop
-    #[test]
-    fn completes_inherited_methods_with_cycle() {
-        let src = "
-        fn main() {
-            trait Inherited2: Inherited1 {}
-            trait Inherited1: Inherited2 {
-                fn inherited(&self);
-            }
-            trait Trait: Inherited1 {
-                fn method(&self);
-            }
-            fn func<T: Trait>(arg: &T) {
-                arg.inheri~
-            }
-        }
-        ";
-        assert_eq!(get_only_completion(src, None).matchstr, "inherited");
-    }
-
-    #[test]
-    fn completes_impled_bounds_for_fnarg() {
-        let src = "
-        fn main() {
-            struct St<T>(T);
-            trait Trait: Sized {
-                fn method(&self);
-            }
-            impl<T: Trait> St<T> {
-                fn new(t: T) -> Self {
-                    t.met~
-                }
-            }
-        }
-        ";
-        assert_eq!(get_only_completion(src, None).matchstr, "method");
-    }
-
-    #[test]
-    fn completes_impled_bounds_for_self() {
-        let src = "
-        fn main() {
-            trait Trait: Sized {
-                fn method(&self);
-            }
-            impl<T: Trait> Clone for T {
-                fn f(self) -> Self {
-                    self.me~
-                }
-            }
-        }
-        ";
-        assert_eq!(get_only_completion(src, None).matchstr, "method");
-    }
-
-    #[test]
-    fn completes_impled_bounds_for_self_field() {
-        let src = "
-        fn main() {
-            struct St<T> {
-                mem1: String,
-                mem2: T,
-            }
-            trait Trait: Sized {
-                fn method(&self);
-            }
-            impl<T: Trait> St<T> {
-                fn f(self) -> Self {
-                    self.mem2.m~
-                }
-            }
-        }
-        ";
-        assert_eq!(get_only_completion(src, None).matchstr, "method");
-    }
-
-    #[test]
-    fn completes_impled_bounds_for_ref_self_field() {
-        let src = "
-        fn main() {
-            struct St<'a, T> {
-                mem1: String,
-                mem2: &'a T,
-            }
-            trait Trait: Sized {
-                fn method(&self);
-            }
-            impl<T: Trait> St<T> {
-                fn f(self) -> Self {
-                    self.mem2.m~
-                }
-            }
-        }
-        ";
-        assert_eq!(get_only_completion(src, None).matchstr, "method");
-    }
-}
-
 #[test]
 fn recursive_glob_depth3() {
     let src = "
@@ -4359,94 +4176,6 @@ fn completes_trait_method_only_once() {
 }
 
 #[test]
-fn completes_methods_for_for_arg() {
-    let src = "
-    fn main() {
-        let v: Vec<String> = vec![];
-        for s in v {
-            s.ca~
-        }
-    }
-    ";
-    let got = get_only_completion(src, None);
-    assert_eq!(got.matchstr, "capacity");
-}
-
-#[test]
-fn completes_methods_for_ref_for_arg() {
-    let src = "
-    #[derive(Copy, Clone)]
-    struct Usize(usize);
-    impl Usize { fn method(&self) {} }
-    fn main() {
-        let v: Vec<&Usize> = vec![];
-        for &u in v {
-            u.me~
-        }
-    }
-    ";
-    let got = get_only_completion(src, None);
-    assert_eq!(got.matchstr, "method");
-}
-
-#[test]
-fn completes_methods_for_tupled_for_arg() {
-    let src = "
-    fn main() {
-        let v: Vec<(String, usize)> = vec![];
-        for (s, i) in v {
-            s.ca~
-        }
-    }
-    ";
-    let got = get_only_completion(src, None);
-    assert_eq!(got.matchstr, "capacity");
-}
-
-#[test]
-fn completes_methods_for_tupls_for_arg() {
-    let src = "
-    fn main() {
-        struct St(String);
-        let v: Vec<St> = vec![];
-        for St(s) in v {
-            s.cap~
-        }
-    }
-    ";
-    let got = get_only_completion(src, None);
-    assert_eq!(got.matchstr, "capacity");
-}
-
-#[test]
-fn completes_methods_for_complex_tupled_for_arg() {
-    let src = "
-    fn main() {
-        struct St(String);
-        let v: Vec<(String, St, usize)> = vec![];
-        for (s1, St(s2), i) in v {
-            s2.ca~
-        }
-    }
-    ";
-    let got = get_only_completion(src, None);
-    assert_eq!(got.matchstr, "capacity");
-    let src = "
-    #[derive(Copy, Clone)]
-    struct Usize(usize);
-    impl Usize { fn method(&self) {} }
-    fn main() {
-        let v: Vec<(String, &Usize)> = vec![];
-        for (s, &u) in v {
-            u.me~
-        }
-    }
-    ";
-    let got = get_only_completion(src, None);
-    assert_eq!(got.matchstr, "method");
-}
-
-#[test]
 fn completes_methods_vec_macro() {
     let src = "
     fn main() {
@@ -4465,72 +4194,6 @@ fn completes_trait_methods_in_path() {
 ";
     let got = get_only_completion(src, None);
     assert_eq!(got.matchstr, "default");
-}
-
-mod completes_methods_for_if_let {
-    use super::*;
-    #[test]
-    fn enum_with_type_annotation() {
-        let src = "
-    fn main() {
-        let s: Option<String> = None;
-        if let Some(s) = s{
-           s.capa~
-        }
-    }
-";
-        let got = get_only_completion(src, None);
-        assert_eq!(got.matchstr, "capacity");
-    }
-
-    #[test]
-    fn enum_variant_with_path() {
-        let src = "
-    fn main() {
-        let s = Some(String::new());
-        if let Some(s) = s{
-           s.capa~
-        }
-    }
-";
-        let got = get_only_completion(src, None);
-        assert_eq!(got.matchstr, "capacity");
-    }
-
-    #[test]
-    fn enum_variant_with_tuple() {
-        let src = "
-    fn main() {
-        let s = Some((String::new(), 0u32));
-        if let Some((s, _)) = s{
-           s.capa~
-        }
-    }
-";
-        let got = get_only_completion(src, None);
-        assert_eq!(got.matchstr, "capacity");
-    }
-}
-
-// blocked by #946
-#[test]
-#[ignore]
-fn completes_methods_for_struct_for_arg() {
-    let src = "
-    fn main() {
-        struct St {
-            name: String,
-            num: usize,
-            __guard: (),
-        }
-        let v: Vec<St> = vec![];
-        for St { name, num: n, .. } in v {
-            name.cap~
-        }
-    }
-    ";
-    let got = get_only_completion(src, None);
-    assert_eq!(got.matchstr, "capacity");
 }
 
 // Experimental featrue
